% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_sr}
\alias{download_sr}
\title{Download Species Richness Data from Williams et al. 2024}
\usage{
download_sr(taxon)
}
\arguments{
\item{taxon}{(bird, mamm) the taxon for which data are downloaded}
}
\value{
the path to the downloaded file
}
\description{
Download Species Richness Data from Williams et al. 2024
}
