% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_convert.R
\name{pdf_to_pngs}
\alias{pdf_to_pngs}
\title{Convert a PDF to Several PNGs (One Per Page)}
\usage{
pdf_to_pngs(pdf_path = "references/williams_natcom_2024.pdf")
}
\arguments{
\item{pdf_path}{the path to the pdf file to convert}
}
\value{
the paths to the png files created (in outputs/)
}
\description{
Convert a PDF to Several PNGs (One Per Page)
}
