% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_convert.R
\name{pngs_to_panel}
\alias{pngs_to_panel}
\title{Convert Several PNGs to a PNG panel}
\usage{
pngs_to_panel(png_paths)
}
\arguments{
\item{png_paths}{the paths to the png files to convert}
}
\value{
the path to the png panel created
}
\description{
Convert Several PNGs to a PNG panel
}
